solutions/solution-00/requirements.txt
flask
flask-cors
flask-MySQLAlchemy

gunicorn
MySQLAlchemy
python-dotenv: 1.0.1